requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
requests==2.31.0
boto3==1.29.7
snowflake-connector-python==3.4.1
python-dotenv==1.0.0
backoff==2.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
ruff==0.1.8
